RK4 bugs solved, tests continue. Hardi Tracking GUI change.

+++ branches/myusufoglu/plugins/stable/HARDIFiberTracking/HARDIdeterministicTracker.h
@@ -152,16 +152,16 @@ class HARDIdeterministicTracker
 			@param pointList	List of fiber points */
 		void calculateFiberDS(int direction, std::vector<HARDIstreamlinePoint> * pointList, std::vector<double*> &anglesArray, vtkIntArray * trianglesArray,int numberOfIterations, bool CLEANMAXIMA, double TRESHOLD);
 		
-		double *findFunctionValue(int threshold, std::vector<double*> &anglesArray, double *weights,  vtkIntArray *trianglesArray, std::vector<int> &meshPtIndexList, std::vector<int> &maxima);
+		double *findFunctionValue(double threshold, std::vector<double*> &anglesArray, double *weights,  vtkIntArray *trianglesArray, std::vector<int> &meshPtIndexList, std::vector<int> &maxima);
 		
-		void findFunctionValueUsingMaximaFile(int threshold, std::vector<double*> &anglesArray, double *weights,  vtkIntArray *trianglesArray, std::vector<int> &meshPtIndexList, std::vector<int> &maxima,double dotLimit=1.0, double *interpolatedVector=nullptr,   int loopAngleSingleCompareOrAverage=0, int loopAngleSelectMaximaCombinationType=0);
+		void findFunctionValueUsingMaximaFile(double threshold, std::vector<double*> &anglesArray, double *weights,  vtkIntArray *trianglesArray, std::vector<int> &meshPtIndexList, std::vector<int> &maxima,double dotLimit=1.0, double *interpolatedVector=nullptr,   int loopAngleSingleCompareOrAverage=0, int loopAngleSelectMaximaCombinationType=0);
 		
 
 	 	
-		void findFunctionValueAtPointUsingMaximaFile(double pos[3],vtkCell * currentCell, vtkIdType currentCellId, int threshold, std::vector<double*> &anglesArray,    vtkIntArray *trianglesArray, std::vector<int> &meshPtIndexList, std::vector<int> &maxima, double *interpolatedVector=nullptr);
+		void findFunctionValueAtPointUsingMaximaFile(double pos[3],vtkCell * currentCell, vtkIdType currentCellId, double threshold, std::vector<double*> &anglesArray,    vtkIntArray *trianglesArray, std::vector<int> &meshPtIndexList, std::vector<int> &maxima, double *interpolatedVector=nullptr);
 
 
-		double * findRK4DeltaX(double pos[3],vtkCell * currentCell, vtkIdType currentCellId, int threshold, std::vector<double*> &anglesArray, double *weights,  vtkIntArray *trianglesArray, std::vector<int> &meshPtIndexList, std::vector<int> &maxima);
+		double * findRK4DeltaX(double pos[3],vtkCell * currentCell, vtkIdType currentCellId, double threshold, std::vector<double*> &anglesArray, double *weights,  vtkIntArray *trianglesArray, std::vector<int> &meshPtIndexList, std::vector<int> &maxima);
 
 		void findMax2( std::vector<double> &array, std::vector<double> &maxima, std::vector<double*> &maximaunitvectors, std::vector<double *> &anglesReturn);
 		/** Sets the unit vectors for this class. */
